+++ commands/commands.go
@@ -12,11 +12,17 @@ import (
 	"golang.org/x/term"
 )
 
+// TODO:
+// 		1. Infos should be write to a file via `-output` flag and user can be provide custom path and file name like this: /path/to/my/file.txt
+// 		2. Wipe functionality, should support go routines, like enc and dec commands.
+// 		3. Update Go.mod, changelog, readme, build.sh (change to kilid not kld), thats it for now.
+
 func Init(kld *engine.Kilid) *cli.Command {
 	app := &cli.Command{
-		Name:  "kld",
+		Name:  "kilid",
 		Usage: "Encrypt & Decrypt your files with Kilid.",
 		Commands: []*cli.Command{
+
 			//ENC
 			{
 				Name:    "enc",
@@ -33,6 +39,10 @@ func Init(kld *engine.Kilid) *cli.Command {
 						Aliases: []string{"y"},
 						Usage:   "Answer 'yes' to all prompts, [!]: existing files will be overwritten",
 					},
+					&cli.BoolFlag{
+						Name:  "wipe",
+						Usage: "Shreds source file to prevent recovery",
+					},
 				},
 				Arguments: []cli.Argument{&cli.StringArgs{
 					Name: "files",
@@ -44,9 +54,14 @@ func Init(kld *engine.Kilid) *cli.Command {
 					var confirmPassword string
 					var hint string
 					files := cmd.StringArgs("files")
-					deleteSource := cmd.Bool("delete")
+					deleteSrc := cmd.Bool("delete")
+					wipeSrc := cmd.Bool("wipe")
 					yesAll := cmd.Bool("yes")
 					scanner := bufio.NewScanner(os.Stdin)
+
+					if deleteSrc && wipeSrc {
+						return fmt.Errorf("you cannot pass 'delete source'(-delete) and 'wipe source'(-wipe) at the same time, use one of them")
+					}
 
 					print("Password: ")
 					b, err := term.ReadPassword(int(os.Stdin.Fd()))
@@ -70,7 +85,7 @@ func Init(kld *engine.Kilid) *cli.Command {
 					scanner.Scan()
 					hint = scanner.Text()
 					fmt.Println()
-					return handlers.HandleEncryption(kld, files, password, hint, deleteSource, yesAll)
+					return handlers.HandleEncryption(kld, files, password, hint, wipeSrc, deleteSrc, yesAll)
 				},
 			},
 
@@ -100,7 +115,7 @@ func Init(kld *engine.Kilid) *cli.Command {
 				Action: func(ctx context.Context, cmd *cli.Command) error {
 					var password string
 					files := cmd.StringArgs("files")
-					deleteSource := cmd.Bool("delete")
+					deleteSrc := cmd.Bool("delete")
 					yesAll := cmd.Bool("yes")
 
 					print("Password: ")
@@ -110,7 +125,7 @@ func Init(kld *engine.Kilid) *cli.Command {
 					}
 					password = string(bytePassword)
 					fmt.Println()
-					return handlers.HandleDecryption(kld, files, password, deleteSource, yesAll)
+					return handlers.HandleDecryption(kld, files, password, deleteSrc, yesAll)
 				},
 			},
 
@@ -125,18 +140,32 @@ func Init(kld *engine.Kilid) *cli.Command {
 						Max:  -1,
 						Min:  1,
 					}},
-
+				Flags: []cli.Flag{
+					&cli.StringFlag{
+						Name:  "output",
+						Usage: "path to save info data, example: --output=folder/to/info.txt",
+					},
+				},
 				Action: func(ctx context.Context, cmd *cli.Command) error {
-					return handlers.HandleInfo(kld, cmd.StringArgs("files"))
+					if cmd.IsSet("output") {
+						fmt.Println("FROM FLAG")
+						output := cmd.StringArg("output")
+						fmt.Println("output is SET, value: ", output)
+						fmt.Println(&output)
+						adr := &output
+						fmt.Println("value of addr", adr, "is: ", *adr)
+						os.Exit(1)
+						return handlers.HandleInfo(kld, cmd.StringArgs("files"), &output)
+					}
+					return handlers.HandleInfo(kld, cmd.StringArgs("files"), nil)
 				},
 			},
 
-			// VERSION
 			{
 				Name:    "version",
 				Aliases: []string{"v"},
 				Usage:   "Show Kilid version",
-				Action: func(ctx context.Context, c *cli.Command) error {
+				Action: func(ctx context.Context, cmd *cli.Command) error {
 					fmt.Println("Version", kld.Version)
 					return nil
 				},