+++ engine/engine.go
@@ -19,26 +19,23 @@ import (
 )
 
 type Kilid struct {
-	Version     string
-	ArgonParams map[string]int
+	Version string
 }
 
 const ChunkSize = 1024 * 1024 // 1MB
 
-func NewKilid(v string, params map[string]int) *Kilid {
+func NewKilid(v string) *Kilid {
 	return &Kilid{
-		Version:     v,
-		ArgonParams: params,
+		Version: v,
 	}
 }
 
 type MetaData struct {
-	OriginalExtension string         `json:"ext"`
-	PasswordHint      string         `json:"hint"`
-	Date              string         `json:"date"`
-	Version           string         `json:"version"`
-	Salt              string         `json:"salt"`
-	ArgonParams       map[string]int `json:"argon_params"`
+	OriginalExtension string `json:"ext"`
+	PasswordHint      string `json:"hint"`
+	Date              string `json:"date"`
+	Version           string `json:"version"`
+	Salt              string `json:"salt"`
 }
 
 func generateSalt() ([]byte, error) {
@@ -49,8 +46,8 @@ func generateSalt() ([]byte, error) {
 	return salt, nil
 }
 
-func deriveKey(password string, salt []byte, params map[string]int) []byte {
-	return argon2.IDKey([]byte(password), salt, uint32(params["iterations"]), uint32(params["memory"]), uint8(params["threads"]), uint32(params["keyLen"]))
+func deriveKey(password string, salt []byte) []byte {
+	return argon2.IDKey([]byte(password), salt, 4, 256*1024, 4, 32)
 }
 
 func (kld *Kilid) EncryptFile(file string, password string, hint string, deleteSource bool, yesAll bool, onProgress func(int)) error {
@@ -77,7 +74,6 @@ func (kld *Kilid) EncryptFile(file string, password string, hint string, deleteS
 	md.PasswordHint = hint
 	md.Date = time.Now().Format("2006-01-02 15:04:05")
 	md.Version = kld.Version
-	md.ArgonParams = kld.ArgonParams
 	md.Salt = hex.EncodeToString(salt)
 
 	b, err := json.Marshal(md)
@@ -91,7 +87,7 @@ func (kld *Kilid) EncryptFile(file string, password string, hint string, deleteS
 		return fmt.Errorf("failed to write metadata delimiter: %w", err)
 	}
 
-	key := deriveKey(password, salt, kld.ArgonParams)
+	key := deriveKey(password, salt)
 	block, err := aes.NewCipher(key)
 	if err != nil {
 		return fmt.Errorf("failed to create block: %w", err)
@@ -144,14 +140,14 @@ func (kld *Kilid) EncryptFile(file string, password string, hint string, deleteS
 	return nil
 }
 
-func (kld *Kilid) DecryptFile(file string, password string, deleteSource bool, yesAll bool, onProgress func(int)) []error {
+func (kld *Kilid) DecryptFile(file string, password string, deleteSource bool, yesAll bool, onProgress func(int)) error {
 	if filepath.Ext(file) != ".kld" {
-		return []error{fmt.Errorf("only .kld files can be decrypted")}
+		return fmt.Errorf("only .kld files can be decrypted")
 	}
 
 	src, err := os.Open(file)
 	if err != nil {
-		return []error{fmt.Errorf("failed to open file: %w", err)}
+		return fmt.Errorf("failed to open file: %w", err)
 	}
 	defer src.Close()
 
@@ -159,35 +155,35 @@ func (kld *Kilid) DecryptFile(file string, password string, deleteSource bool, y
 	r := bufio.NewReader(src)
 	b, err := r.ReadBytes('/')
 	if err != nil {
-		return []error{fmt.Errorf("failed to read metadata: %w", err)}
+		return fmt.Errorf("failed to read metadata: %w", err)
 	}
 	onProgress(len(b))
 	metadataStr := string(b)
 	metadataStr = strings.TrimSuffix(metadataStr, "/")
 
 	if err := json.Unmarshal([]byte(metadataStr), &md); err != nil {
-		return []error{fmt.Errorf("failed to parse metadata: %w", err)}
+		return fmt.Errorf("failed to parse metadata: %w", err)
 	}
 
 	dst, err := os.Create(kld.GetFileName(file) + md.OriginalExtension)
 	if err != nil {
-		return []error{fmt.Errorf("failed to create output file: %w", err)}
+		return fmt.Errorf("failed to create output file: %w", err)
 	}
 	defer dst.Close()
 
 	salt, err := hex.DecodeString(md.Salt)
 	if err != nil {
-		return []error{fmt.Errorf("failed to convert salt: %w", err)}
+		return fmt.Errorf("failed to convert salt: %w", err)
 	}
 
-	block, err := aes.NewCipher(deriveKey(password, salt, md.ArgonParams))
+	block, err := aes.NewCipher(deriveKey(password, salt))
 	if err != nil {
-		return []error{fmt.Errorf("failed to create block: %w", err)}
+		return fmt.Errorf("failed to create block: %w", err)
 	}
 
 	gcm, err := cipher.NewGCM(block)
 	if err != nil {
-		return []error{fmt.Errorf("failed to create GCM counter: %w", err)}
+		return fmt.Errorf("failed to create GCM counter: %w", err)
 	}
 
 	nonceSize := gcm.NonceSize()
@@ -199,30 +195,27 @@ func (kld *Kilid) DecryptFile(file string, password string, deleteSource bool, y
 			break
 		}
 		if err != nil {
-			return []error{fmt.Errorf("failed to read length: %w", err)}
+			return fmt.Errorf("failed to read length: %w", err)
 		}
 
 		nonce := make([]byte, nonceSize)
 		if _, err := io.ReadFull(r, nonce); err != nil {
-			return []error{fmt.Errorf("failed to read nonce: %w", err)}
+			return fmt.Errorf("failed to read nonce: %w", err)
 		}
 
 		ciphertext := make([]byte, length)
 		if _, err := io.ReadFull(r, ciphertext); err != nil {
-			return []error{fmt.Errorf("failed to read ciphertext: %w", err)}
+			return fmt.Errorf("failed to read ciphertext: %w", err)
 		}
 
 		plaindata, err := gcm.Open(nil, nonce, ciphertext, nil)
 		if err != nil {
 			dst.Close()
-			if err := os.Remove(dst.Name()); err != nil {
-				return []error{fmt.Errorf("failed to delete temoporary output file: %w", err), fmt.Errorf("password is wrong or data tampered"), fmt.Errorf("Password Hint: %s", md.PasswordHint)}
-			}
-			return []error{fmt.Errorf("password is wrong or data tampered"), fmt.Errorf("Password Hint: %s", md.PasswordHint)}
+			return fmt.Errorf("password is wrong or data tampered | Password Hint: %q", md.PasswordHint)
 		}
 
 		if _, err := dst.Write(plaindata); err != nil {
-			return []error{fmt.Errorf("failed to save data: %w", err)}
+			return fmt.Errorf("failed to save data: %w", err)
 		}
 
 		bytesConsumed := 4 + nonceSize + int(length)
@@ -231,27 +224,90 @@ func (kld *Kilid) DecryptFile(file string, password string, deleteSource bool, y
 
 	if deleteSource {
 		if err := deleteSourceFile(file); err != nil {
-			return []error{err}
+			return err
 		}
 	}
 
 	return nil
-
 }
 
-func (kld *Kilid) Info(file string, onlyExt bool) (string, error) {
+func (kld *Kilid) WipeFile(file string, onProgress func(int)) error {
+	src, err := os.OpenFile(file, os.O_RDWR, 0644)
+	if err != nil {
+		return fmt.Errorf("failed to open file: %w", err)
+	}
+	defer src.Close()
+
+	stats, err := src.Stat()
+	if err != nil {
+		return fmt.Errorf("failed to get file statistics: %w", err)
+	}
+	fileSize := stats.Size()
+
+	buf := make([]byte, ChunkSize)
+	var totalProcessed int64
+
+	for totalProcessed < fileSize {
+		n, err := src.Read(buf)
+		if n > 0 {
+			scrambledData := make([]byte, n)
+			if _, err = rand.Read(scrambledData); err != nil {
+				return fmt.Errorf("failed to fill scramble data: %w", err)
+			}
+
+			if _, err = src.Seek(int64(-n), io.SeekCurrent); err != nil {
+				return fmt.Errorf("failed to seek file cursor: %w", err)
+			}
+
+			if _, err = src.Write(scrambledData); err != nil {
+				return fmt.Errorf("failed to write scrabled data: %w", err)
+			}
+
+			totalProcessed += int64(n)
+			// onProgress(int(totalProcessed))
+		}
+		if err == io.EOF {
+			break
+		}
+		if err != nil {
+			return fmt.Errorf("failed to read file: %w", err)
+		}
+	}
+
+	return nil
+}
+
+func (kld *Kilid) Info(file string) (MetaData, error) {
+	md, err := getMetadata(file)
+	if err != nil {
+		return MetaData{}, fmt.Errorf("failed to get metadata: %w", err)
+	}
+
+	return md, nil
+}
+
+func (kld *Kilid) GetFileRealExt(file string) (string, error) {
+	md, err := getMetadata(file)
+	if err != nil {
+		return "", fmt.Errorf("failed to get file extension: %w", err)
+	}
+
+	return md.OriginalExtension, nil
+}
+
+func getMetadata(file string) (MetaData, error) {
 	if filepath.Ext(file) != ".kld" {
-		return "", fmt.Errorf("only .kld files can be printed")
+		return MetaData{}, fmt.Errorf("only .kld files can be printed")
 	}
 
 	src, err := os.Open(file)
 	if err != nil {
-		return "", fmt.Errorf("failed to open file: %w", err)
+		return MetaData{}, fmt.Errorf("failed to open file: %w", err)
 	}
 	r := bufio.NewReader(src)
 	b, err := r.ReadBytes('/')
 	if err != nil {
-		return "", fmt.Errorf("failed to read metadata: %w", err)
+		return MetaData{}, fmt.Errorf("failed to read metadata: %w", err)
 	}
 	src.Close()
 	mdString := string(b)
@@ -260,20 +316,10 @@ func (kld *Kilid) Info(file string, onlyExt bool) (string, error) {
 	var md MetaData
 
 	if err := json.Unmarshal([]byte(mdString), &md); err != nil {
-		return "", fmt.Errorf("failed to parse metadata: %w", err)
+		return MetaData{}, fmt.Errorf("failed to parse metadata: %w", err)
 	}
 
-	if onlyExt {
-		return md.OriginalExtension, nil
-	}
-
-	fmt.Printf("\n─── File Details: %s ───\n", file)
-	fmt.Printf("%-20s : %s\n", "Original Extension", md.OriginalExtension)
-	fmt.Printf("%-20s : %s\n", "Password Hint", md.PasswordHint)
-	fmt.Printf("%-20s : %s\n", "Encrypted At", md.Date)
-	fmt.Printf("%-20s : %v\n", "Kilid Version", md.Version)
-	fmt.Println(strings.Repeat("─", 45))
-	return "", nil
+	return md, nil
 }
 
 func (kld *Kilid) GetFileName(file string) string {